(lesson16): Include folder restructure
