+++ src/bsp.c
@@ -1,5 +1,5 @@
 #include "../include/defs.h"
-#include "../include/stm32f429xx.h"
+#include "../include/cmsis/stm32f429xx.h"
 
 void assert_failed(char const* file, int line) {
     /**
